refactor(#1917): single coercion engine — emitToNumber Step 2 (behavior-neutral)
Step 2 of the phased coercion dedup (stacked on Step 1 / PR #1960). Adds
emitToNumber(ctx, fctx, valType) to coercion-engine.ts — the consolidated
ToNumber cascade: void→NaN, i64→f64, externref→__unbox_number (js-host) /
coerceType(f64,"number") (standalone), object-ref→coerceType(f64,"number"),
i32→f64, f64 no-op. The externref arm is gated on ctx.standalone EXACTLY (not
noJsHost) to match the migrated Number(x) site byte-for-byte under --target wasi.

Migrated: the Number(x) lowering (calls.ts) — its post-pre-check cascade now
calls emitToNumber. Kept in the caller (each is a source special-case the
engine can't carry): the #2160 Number(arr) array→ToString→StringToNumber
pre-check, the Symbol-throw, and the native-string-ref→__str_to_number
typeIdx-keyed pre-check.

Deferred to a follow-up increment (NOT a missed copy — different ToNumber
policy, folding would REGRESS): the unary +/-/~ arms use coerceType(f64) with
the DEFAULT hint for externref/ref, whereas Number(x) uses "number"/__unbox_number.
Unifying them neutrally needs an explicit hint param + careful coerceType-vs-
__unbox_number equivalence tracing — a separate neutrality analysis.

#2108 ratcheted DOWN: calls.ts 27→26 (the Number(x) __unbox_number use now in
the sanctioned engine). No unsanctioned growth.

Validation: tsc clean; prettier clean; coercion-sites gate OK. Number(arr)
suite 13/13 substantive pass; the 1 bareEmptyNoCrash failure (Number([]) bare
never[] literal) reproduces IDENTICALLY on the Step-1 base worktree (verified)
— pre-existing, not a regression.
